--- controller/tracepipeline/setup_test.go
@@ -1,90 +1,0 @@
-package tracepipeline
-
-import (
-	"testing"
-
-	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
-	"github.com/stretchr/testify/require"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-)
-
-func TestMapSecret(t *testing.T) {
-	tests := []struct {
-		password         telemetryv1alpha1.ValueType
-		secret           corev1.Secret
-		summary          string
-		expectedRequests []reconcile.Request
-	}{
-		{
-			summary: "map secret referenced by pipeline",
-			password: telemetryv1alpha1.ValueType{
-				ValueFrom: &telemetryv1alpha1.ValueFromSource{
-					SecretKeyRef: &telemetryv1alpha1.SecretKeyRef{
-						Name:      "basic-auth-credentials",
-						Namespace: "default",
-						Key:       "password",
-					},
-				},
-			},
-			secret: corev1.Secret{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "basic-auth-credentials",
-					Namespace: "default",
-				},
-			},
-			expectedRequests: []reconcile.Request{{NamespacedName: types.NamespacedName{Name: "dummy"}}},
-		},
-		{
-			summary: "map unused secret",
-			password: telemetryv1alpha1.ValueType{
-				Value: "qwerty",
-			},
-			secret: corev1.Secret{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "basic-auth-credentials",
-					Namespace: "default",
-				},
-			},
-			expectedRequests: []reconcile.Request{},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.summary, func(t *testing.T) {
-			tracePipeline := &telemetryv1alpha1.TracePipeline{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "dummy",
-				},
-				Spec: telemetryv1alpha1.TracePipelineSpec{
-					Output: telemetryv1alpha1.TracePipelineOutput{
-						Otlp: &telemetryv1alpha1.OtlpOutput{
-							Endpoint: telemetryv1alpha1.ValueType{Value: "localhost"},
-							Authentication: &telemetryv1alpha1.AuthenticationOptions{
-								Basic: &telemetryv1alpha1.BasicAuthOptions{
-									User:     telemetryv1alpha1.ValueType{Value: "admin"},
-									Password: tc.password,
-								},
-							},
-						},
-					},
-				},
-			}
-
-			scheme := runtime.NewScheme()
-			_ = clientgoscheme.AddToScheme(scheme)
-			_ = telemetryv1alpha1.AddToScheme(scheme)
-			fakeClient := fake.NewClientBuilder().WithScheme(scheme).WithObjects(tracePipeline).Build()
-			sut := Reconciler{Client: fakeClient}
-
-			actualRequests := sut.mapSecret(&tc.secret)
-
-			require.ElementsMatch(t, tc.expectedRequests, actualRequests)
-		})
-	}
-}
