+++ internal/reconciler/logparser/mocks/daemon_set_annotator.go
