+++ internal/reconciler/logparser/mocks/daemon_set_prober.go
