+++ internal/reconciler/logparser/status_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
-	"github.com/kyma-project/telemetry-manager/controller/logparser/mocks"
+	"github.com/kyma-project/telemetry-manager/internal/reconciler/logparser/mocks"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
