+++ internal/reconciler/logpipeline/mocks/daemon_set_annotator.go
