+++ internal/reconciler/logpipeline/mocks/daemon_set_prober.go
