+++ internal/reconciler/logpipeline/status.go
