+++ internal/reconciler/logpipeline/sync.go
