+++ internal/reconciler/tracepipeline/mocks/configmap_prober.go
