+++ internal/reconciler/tracepipeline/mocks/deployment_prober.go
