+++ internal/reconciler/tracepipeline/reconciler.go
@@ -20,22 +20,21 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"strings"
+
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
 	"github.com/kyma-project/telemetry-manager/internal/configchecksum"
 	utils "github.com/kyma-project/telemetry-manager/internal/kubernetes"
 	"github.com/kyma-project/telemetry-manager/internal/overrides"
 	commonresources "github.com/kyma-project/telemetry-manager/internal/resources/common"
 	corev1 "k8s.io/api/core/v1"
-	"strings"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
 )
 
 type Config struct {
@@ -73,42 +72,15 @@ type Reconciler struct {
 	overridesHandler overrides.GlobalConfigHandler
 }
 
-func NewReconciler(client client.Client, config Config, prober DeploymentProber, handler *overrides.Handler) *Reconciler {
+func NewReconciler(client client.Client, config Config, prober DeploymentProber) *Reconciler {
 	var r Reconciler
 	r.Client = client
 	r.config = config
 	r.prober = prober
-	r.overridesHandler = handler
 	return &r
 }
 
-func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (reconcileResult ctrl.Result, reconcileErr error) {
-	log := logf.FromContext(ctx)
-
-	log.V(1).Info("Reconciliation triggered")
-
-	overrideConfig, err := r.overridesHandler.UpdateOverrideConfig(ctx, r.config.OverrideConfigMap)
-	if err != nil {
-		return ctrl.Result{}, err
-	}
-
-	if err := r.overridesHandler.CheckGlobalConfig(overrideConfig.Global); err != nil {
-		return ctrl.Result{}, err
-	}
-	if overrideConfig.Tracing.Paused {
-		log.V(1).Info("Skipping reconciliation of tracepipeline as reconciliation is paused")
-		return ctrl.Result{}, nil
-	}
-
-	var tracePipeline telemetryv1alpha1.TracePipeline
-	if err := r.Get(ctx, req.NamespacedName, &tracePipeline); err != nil {
-		return ctrl.Result{}, client.IgnoreNotFound(err)
-	}
-
-	return ctrl.Result{}, r.doReconcile(ctx, &tracePipeline)
-}
-
-func (r *Reconciler) doReconcile(ctx context.Context, pipeline *telemetryv1alpha1.TracePipeline) error {
+func (r *Reconciler) DoReconcile(ctx context.Context, pipeline *telemetryv1alpha1.TracePipeline) error {
 	var err error
 	lockAcquired := true
