+++ internal/reconciler/tracepipeline/secret_refs.go
@@ -114,3 +114,30 @@ func appendOutputFieldIfHasSecretRef(fields []fieldDescriptor, pipelineName stri
 
 	return fields
 }
+
+func ContainsAnyRefToSecret(pipeline *telemetryv1alpha1.TracePipeline, secret *corev1.Secret) bool {
+	secretName := types.NamespacedName{Namespace: secret.Namespace, Name: secret.Name}
+	if pipeline.Spec.Output.Otlp.Endpoint.IsDefined() &&
+		referencesSecret(pipeline.Spec.Output.Otlp.Endpoint, secretName) {
+		return true
+	}
+
+	if pipeline.Spec.Output.Otlp == nil ||
+		pipeline.Spec.Output.Otlp.Authentication == nil ||
+		pipeline.Spec.Output.Otlp.Authentication.Basic == nil ||
+		!pipeline.Spec.Output.Otlp.Authentication.Basic.IsDefined() {
+		return false
+	}
+
+	auth := pipeline.Spec.Output.Otlp.Authentication.Basic
+
+	return referencesSecret(auth.User, secretName) || referencesSecret(auth.Password, secretName)
+}
+
+func referencesSecret(valueType telemetryv1alpha1.ValueType, secretName types.NamespacedName) bool {
+	if valueType.Value == "" && valueType.ValueFrom != nil && valueType.ValueFrom.IsSecretKeyRef() {
+		return valueType.ValueFrom.SecretKeyRef.NamespacedName() == secretName
+	}
+
+	return false
+}
